Knowledge from NERACOOS
Adding many more of the standards that we’ve seen on NERACOOS servers, 
and in the process realizing that there are a few other ways and bits of
data we may want to store with knowledge.

`sibling_standards` is for standards that are most commonly seen 
together, as usually they come from the same suite of sensors.

`extra_attrs` is additional attributes like NDBC names that we may want 
to add to a dataset.